+++ media/cdm/cbcs_decryptor_unittest.cc
@@ -148,7 +148,8 @@ TEST_F(CbcsDecryptorTest, OneBlock) {
   DCHECK_EQ(kBlockSize, encrypted_block.size());
 
   // Only 1 subsample, all encrypted data.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(
       encrypted_block, iv_, subsamples, EncryptionPattern(1, 9));
@@ -160,7 +161,8 @@ TEST_F(CbcsDecryptorTest, AdditionalData) {
   DCHECK_EQ(kBlockSize, encrypted_block.size());
 
   // Only 1 subsample, all encrypted data.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(
       encrypted_block, iv_, subsamples, EncryptionPattern(1, 9));
@@ -190,7 +192,8 @@ TEST_F(CbcsDecryptorTest, DifferentPattern) {
   DCHECK_EQ(kBlockSize, encrypted_block.size());
 
   // Only 1 subsample, all encrypted data.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(
       encrypted_block, iv_, subsamples, EncryptionPattern(1, 0));
@@ -202,7 +205,8 @@ TEST_F(CbcsDecryptorTest, EmptyPattern) {
   DCHECK_EQ(kBlockSize, encrypted_block.size());
 
   // Only 1 subsample, all encrypted data.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   // Pattern 0:0 treats the buffer as all encrypted.
   auto encrypted_buffer = CreateEncryptedBuffer(
@@ -215,7 +219,8 @@ TEST_F(CbcsDecryptorTest, PatternTooLarge) {
   DCHECK_EQ(kBlockSize, encrypted_block.size());
 
   // Only 1 subsample, all encrypted data.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   // Pattern 100:0 is too large, so decryption will fail.
   auto encrypted_buffer = CreateEncryptedBuffer(
@@ -238,7 +243,8 @@ TEST_F(CbcsDecryptorTest, BadSubsamples) {
   auto encrypted_block = Encrypt(one_block_, *key_, iv_);
 
   // Subsample size > data size.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size() + 1}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size() + 1)}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(
       encrypted_block, iv_, subsamples, EncryptionPattern(1, 0));
@@ -248,7 +254,8 @@ TEST_F(CbcsDecryptorTest, BadSubsamples) {
 TEST_F(CbcsDecryptorTest, InvalidIv) {
   auto encrypted_block = Encrypt(one_block_, *key_, iv_);
 
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   // Use an invalid IV for decryption. Call should succeed, but return
   // something other than the original data.
@@ -261,7 +268,8 @@ TEST_F(CbcsDecryptorTest, InvalidIv) {
 TEST_F(CbcsDecryptorTest, InvalidKey) {
   auto encrypted_block = Encrypt(one_block_, *key_, iv_);
 
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   // Use a different key for decryption. Call should succeed, but return
   // something other than the original data.
@@ -275,7 +283,8 @@ TEST_F(CbcsDecryptorTest, InvalidKey) {
 TEST_F(CbcsDecryptorTest, PartialBlock) {
   // Only 1 subsample, all "encrypted" data. However, as it's not a full block,
   // it will be treated as unencrypted.
-  std::vector<SubsampleEntry> subsamples = {{0, partial_block_.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(partial_block_.size())}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(partial_block_, iv_, subsamples,
                                                 EncryptionPattern(1, 0));
@@ -290,7 +299,8 @@ TEST_F(CbcsDecryptorTest, SingleBlockWithExtraData) {
   auto expected_result = Combine({one_block_, partial_block_});
 
   // Only 1 subsample, all "encrypted" data.
-  std::vector<SubsampleEntry> subsamples = {{0, encrypted_block.size()}};
+  std::vector<SubsampleEntry> subsamples = {
+      {0, static_cast<uint32_t>(encrypted_block.size())}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(
       encrypted_block, iv_, subsamples, EncryptionPattern(1, 0));
@@ -299,7 +309,8 @@ TEST_F(CbcsDecryptorTest, SingleBlockWithExtraData) {
 
 TEST_F(CbcsDecryptorTest, SkipBlock) {
   // Only 1 subsample, but all unencrypted data.
-  std::vector<SubsampleEntry> subsamples = {{one_block_.size(), 0}};
+  std::vector<SubsampleEntry> subsamples = {
+      {static_cast<uint32_t>(one_block_.size()), 0}};
 
   auto encrypted_buffer = CreateEncryptedBuffer(one_block_, iv_, subsamples,
                                                 EncryptionPattern(1, 0));
