+++ media/mojo/common/mojo_shared_buffer_video_frame.cc
@@ -61,7 +61,8 @@ MojoSharedBufferVideoFrame::CreateDefaultForTesting(
     return Create(
         format, coded_size, visible_rect, dimensions, std::move(handle),
         allocation_size,
-        {0 /* y_offset */, coded_size.GetArea(), coded_size.GetArea() * 5 / 4},
+        {0 /* y_offset */, static_cast<uint32_t>(coded_size.GetArea()),
+         static_cast<uint32_t>(coded_size.GetArea() * 5 / 4)},
         {coded_size.width(), coded_size.width() / 2, coded_size.width() / 2},
         timestamp);
   } else {
@@ -71,10 +72,11 @@ MojoSharedBufferVideoFrame::CreateDefaultForTesting(
     // as follows:
     //  - Yplane, full size (each element represents a 1x1 block)
     //  - UVplane, full width, half height (each pair represents a 2x2 block)
-    return Create(format, coded_size, visible_rect, dimensions,
-                  std::move(handle), allocation_size,
-                  {0 /* y_offset */, coded_size.GetArea()},
-                  {coded_size.width(), coded_size.width()}, timestamp);
+    return Create(
+        format, coded_size, visible_rect, dimensions, std::move(handle),
+        allocation_size,
+        {0 /* y_offset */, static_cast<uint32_t>(coded_size.GetArea())},
+        {coded_size.width(), coded_size.width()}, timestamp);
   }
 }
